% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{chunk}
\alias{chunk}
\title{Oneliner that splits a vector (x) into n chunks
helpful for string formatting when we have discrete text we want to avoid splitting}
\usage{
chunk(x, n)
}
\arguments{
\item{x}{a vector (not-necessarily atomic) to split}

\item{n}{numeric number of bins to split x into}
}
\description{
Oneliner that splits a vector (x) into n chunks
helpful for string formatting when we have discrete text we want to avoid splitting
}
