% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_pretty_labels}
\alias{cut_pretty_labels}
\title{A thin wrapper around Hmisc::cut2 which
relabels the output labels using a format string}
\usage{
cut_pretty_labels(x, fmt = "\%s-\%s", last_val_plus = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector to bin using cut2}

\item{fmt}{A string interpretable by sprintf formatting the breaks}

\item{last_val_plus}{(default: FALSE) Should the final break be an interval or "{max(x)}+"}

\item{...}{additional parameters passed thru to cut2}
}
\description{
A thin wrapper around Hmisc::cut2 which
relabels the output labels using a format string
}
