% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cut_relabel}
\alias{cut_relabel}
\title{Relabel bins from cut2 nicely for plotting
NOTE: Assumes Hmisc::cut2(., oneval=FALSE)}
\usage{
cut_relabel(str, fmt = "\%s - \%s", big_mark = NULL, last_value_gte = FALSE)
}
\arguments{
\item{str}{a character vector of labels from cut2 to re-label}

\item{fmt}{a format string interpretable by sprintf to format the labels with}

\item{last_value_gte}{(boolean) If \code{TRUE}, will output last level with ">=number" notation}
}
\description{
Relabel bins from cut2 nicely for plotting
NOTE: Assumes Hmisc::cut2(., oneval=FALSE)
}
