% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endemic_countries}
\alias{endemic_countries}
\title{Monkeypox Endemic Countries}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{\code{Country}}{character English name for administrative region}
\item{\code{Endemic}}{character "Y"}
\item{\code{iso3code}}{character ISO 3166-1 alpha-3 country code}
}
}
\usage{
endemic_countries
}
\description{
A data frame of countries that are known to be endemic for Monkeypox.
}
\details{
This is used internally for filtering for some tables and viz.
}
\keyword{datasets}
