% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_denoms.R
\name{get_global_denoms}
\alias{get_global_denoms}
\title{Update Global Population Denominators}
\usage{
get_global_denoms(vintage = 2021, country_metadata = NULL)
}
\arguments{
\item{vintage}{(numeric, default: 2021) The year of population projections to use from UN data}

\item{country_metadata}{(data.frame, optional) A data.frame containing country-specific metadata with at least four columns: \link{"iso3code", "iso3code", "state_region", "who_region"}}
}
\value{
Returns a df of 238 rows and 10 columns
}
\description{
Output is available through the package as "global_denoms," but this function can be used to recreate this dataset.

Note: World Health Organization (WHO) region, Department of State (DoS) region, and english country names are handled externally in a CSV file.
}
\section{Note}{

\itemize{
\item \code{country_metadata} overrides metadata contained in package. Your milage may vary.
\item Population updates for Pitcairn Islands are hardcoded and must be pulled manually via CIA factbook unless another source is found.
}
}

\examples{
\dontrun{
# UPDATING global_denoms
global_denoms <- get_global_denoms()
usethis::use_data(global_denoms, overwrite = TRUE)
write_csv(global_denoms, "inst/extdata/global_denoms.csv")
}
}
\seealso{
\link{global_denoms} for more complete data documentation
}
