% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_accessors.R
\name{get_mpx_cases}
\alias{get_mpx_cases}
\title{Pull MPX Case Data}
\usage{
get_mpx_cases(path, connection = NULL)
}
\arguments{
\item{path}{character string path to MPX data}

\item{connection}{(optional) an \code{spoConnection} object}

\item{...}{additional arguments handled by other methods (not yet implemented)}
}
\description{
Pull MPX Case Data
}
