% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_map}
\alias{global_map}
\title{Global Map Shapefile}
\format{
A Simple Feature Collection with 174 rows and 4 variables:
\describe{
\item{\code{TYPE}}{character one of ("Sovereign country", "Country", "Dependency", "Disputed", NA)}
\item{\code{ADMIN}}{character English name for administrative region}
\item{\code{iso3code}}{character ISO 3166-1 alpha-3 country code}
\item{\code{geometry}}{list sf geometry column}
}
}
\usage{
global_map
}
\description{
A standard shapefile we use for mapping tasks
}
\details{
Liable to update, possibly (likely) incorrect spherical geometry.
}
\keyword{datasets}
