% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mpx_table_cases}
\alias{mpx_table_cases}
\title{Top 10 Table Summaries}
\usage{
mpx_table_cases(df)
}
\arguments{
\item{df}{A Monkeypox data.frame with at least \code{iso3code} and \code{cases} columns}
}
\description{
Top 10 Table Summaries
}
