% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{mpx_table_newcases}
\alias{mpx_table_newcases}
\title{Top 10 Table Summaries}
\usage{
mpx_table_newcases(df, date_since)
}
\arguments{
\item{df}{A Monkeypox data.frame with at least \code{iso3code} and \code{cases} columns}

\item{date_since}{use date format YYYY-MM-DD}
}
\description{
Top 10 Table Summaries
}
