% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{spoConnection}
\alias{spoConnection}
\title{An R6 Class to interface CDC Sharepoint sites}
\description{
An R6 Class to interface CDC Sharepoint sites

An R6 Class to interface CDC Sharepoint sites
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{spoConnection$new()}}
\item \href{#method-read_file}{\code{spoConnection$read_file()}}
\item \href{#method-list_files}{\code{spoConnection$list_files()}}
\item \href{#method-get_list}{\code{spoConnection$get_list()}}
\item \href{#method-write_file}{\code{spoConnection$write_file()}}
\item \href{#method-clone}{\code{spoConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Sharepoint Site connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spoConnection$new(tenant, client_id, client_secret, teams_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tenant}}{(character) An Azure tenant string}

\item{\code{client_id}}{(character) A Client ID for the App Registration}

\item{\code{client_secret}}{(character) A Client Secret for the App Registration}

\item{\code{teams_name}}{(character) Either a MS Teams name or Sharepoint site name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{spoConnection} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_file"></a>}}
\if{latex}{\out{\hypertarget{method-read_file}{}}}
\subsection{Method \code{read_file()}}{
Read in a file from Sharepoint Site
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spoConnection$read_file(relative_url, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relative_url}}{File path relative to the "Documents" folder}

\item{\code{...}}{other named parameters passed to a read in function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If Excel or CSV file, a data frame. All other types will return error
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_files"></a>}}
\if{latex}{\out{\hypertarget{method-list_files}{}}}
\subsection{Method \code{list_files()}}{
List files in a Teams / Sharepoint Folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spoConnection$list_files(relative_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relative_url}}{(character) a folder URL relative to the "Shared Documents" folder}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of files present in the folder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_list}{}}}
\subsection{Method \code{get_list()}}{
Pull a Sharepoint List from an MS Teams/Sharepoint Site
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spoConnection$get_list(list_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list_name}}{(character) Name of the Sharepoint list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Possibly a data.frame of the sharepoint list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_file"></a>}}
\if{latex}{\out{\hypertarget{method-write_file}{}}}
\subsection{Method \code{write_file()}}{
Write a local file to a MS Teams/Sharepoint Site
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spoConnection$write_file(x, relative_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(character) Local path to the file to write}

\item{\code{relative_url}}{(character) Destination path where the file should be written to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
URL path linking to the newly created file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spoConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
