% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_accessors.R
\name{get_mpx_cases}
\alias{get_mpx_cases}
\title{Pull MPX Case Data}
\usage{
get_mpx_cases(path, connection = NULL)
}
\arguments{
\item{path}{character string path to MPX data}

\item{connection}{(optional) an \code{spoConnection} object}
}
\value{
A data frame with n rows and 4 variables:
\itemize{
\item{\code{Country}}{ character English name for administrative region}
\item{\code{date}}{ date reporting date}
\item{\code{cases}}{ integer cumulative monkeypox cases}
\item{\code{iso3code}}{ character ISO 3166-1 alpha-3 country code}
}
}
\description{
Pull MPX Case Data
}
